refactor(State): MasterMind::State with State design pattern
The previous implementation of MasterMind::State was mess a and confusing. It required moving around logic that was all nested into a single method with a conditional that was getting larger. I've moved each 'state' into there own singleton classes that implement an update method. Each class is only concerned on what conditions need to be met for the transition to the next logical state.